API BREAK: dict -> namemedtuple for infos
The API needs to export informations to the caller:
stream informations, codec parameters and stuff.
Most of them are read-only, some are not (encoder params).

We need a general purpose machinery to export those infos
using a familiar, friendly and fast dot notation.
Using plain dicts feels clunky and akward, and frozendicts
are just an helpful bandaid fix.